Moved logger util
