+++ src/tasks/start/backend/backend.ts
@@ -6,7 +6,7 @@ import { createProxy, updateProxy } from './proxy'
 import { resolveRepo, majorBumpRepo } from './repo'
 import { setAllPermissionsOpenly } from './permissions'
 import { KernelInstance, RepoInstance } from '~/typechain'
-import { logBack } from '../utils/logger'
+import { logBack } from '../../../ui/logger'
 import { readArapp } from '../../../utils/arappUtils'
 import { AragonConfig, AragonConfigHooks } from '~/src/types'
 import { TASK_COMPILE } from '../../task-names'
