+++ src/tasks/start/frontend/frontend.ts
@@ -1,6 +1,6 @@
 import chokidar from 'chokidar'
 import { BuidlerRuntimeEnvironment } from '@nomiclabs/buidler/types'
-import { logFront } from '../utils/logger'
+import { logFront } from '../../../ui/logger'
 import { AragonConfig } from '~/src/types'
 import { emitEvent, FRONTEND_STARTED_SERVING } from '../../../events'
 import {
