+++ src/tasks/start/start-task.ts
@@ -2,7 +2,7 @@ import { task } from '@nomiclabs/buidler/config'
 import { BuidlerPluginError } from '@nomiclabs/buidler/plugins'
 import { BuidlerRuntimeEnvironment } from '@nomiclabs/buidler/types'
 import { TASK_START } from '../task-names'
-import { logMain } from './utils/logger'
+import { logMain } from '../../ui/logger'
 import { startBackend } from './backend/backend'
 import { startFrontend } from './frontend/frontend'
 import { AragonConfig } from '~/src/types'
